+++ demo/dapp/public/index.html
@@ -8,8 +8,6 @@
 	<title>Rebase</title>
 
 	<link rel="icon" type="image/png" href="/assets/rebase_icon.svg">
-	<link rel="icon" type="image/png" href="/assets/rebase_icon.svg">
-
 	<link rel="stylesheet" href="/build/bundle.css">
 	<script defer src="/build/bundle.js"></script>
 
@@ -18,18 +16,18 @@
 			font-family: "Basier Square";
 			font-style: normal;
 			font-weight: 400;
-			src: url("./fonts/basier-square-regular.eot");
+			src: url("/fonts/basier-square-regular.eot");
 			/* IE9 Compat Modes */
 			src: local(""),
-				url("./fonts/basier-square-regular.eot?#iefix") format("embedded-opentype"),
+				url("/fonts/basier-square-regular.eot?#iefix") format("embedded-opentype"),
 				/* IE6-IE8 */
-				url("./fonts/basier-square-regular.woff2") format("woff2"),
+				url("/fonts/basier-square-regular.woff2") format("woff2"),
 				/* Super Modern Browsers */
-				url("./fonts/basier-square-regular.woff") format("woff"),
+				url("/fonts/basier-square-regular.woff") format("woff"),
 				/* Modern Browsers */
-				url("./fonts/basier-square-regular.ttf") format("truetype"),
+				url("/fonts/basier-square-regular.ttf") format("truetype"),
 				/* Safari, Android, iOS */
-				url("./fonts/basier-square-regular.svg#Basier Square") format("svg");
+				url("/fonts/basier-square-regular.svg#Basier Square") format("svg");
 			/* Legacy iOS */
 		}
 
@@ -37,18 +35,18 @@
 			font-family: "Basier Square";
 			font-style: normal;
 			font-weight: 600;
-			src: url("./fonts/basier-square-medium.eot");
+			src: url("/fonts/basier-square-medium.eot");
 			/* IE9 Compat Modes */
 			src: local(""),
-				url("./fonts/basier-square-medium.eot?#iefix") format("embedded-opentype"),
+				url("/fonts/basier-square-medium.eot?#iefix") format("embedded-opentype"),
 				/* IE6-IE8 */
-				url("./fonts/basier-square-medium.woff2") format("woff2"),
+				url("/fonts/basier-square-medium.woff2") format("woff2"),
 				/* Super Modern Browsers */
-				url("./fonts/basier-square-medium.woff") format("woff"),
+				url("/fonts/basier-square-medium.woff") format("woff"),
 				/* Modern Browsers */
-				url("./fonts/basier-square-medium.ttf") format("truetype"),
+				url("/fonts/basier-square-medium.ttf") format("truetype"),
 				/* Safari, Android, iOS */
-				url("./fonts/basier-square-medium.svg#Basier Square") format("svg");
+				url("/fonts/basier-square-medium.svg#Basier Square") format("svg");
 			/* Legacy iOS */
 		}
 
@@ -56,18 +54,18 @@
 			font-family: "Basier Square";
 			font-style: normal;
 			font-weight: 700;
-			src: url("./fonts/basier-square-bold.eot");
+			src: url("/fonts/basier-square-bold.eot");
 			/* IE9 Compat Modes */
 			src: local(""),
-				url("./fonts/basier-square-bold.eot?#iefix") format("embedded-opentype"),
+				url("/fonts/basier-square-bold.eot?#iefix") format("embedded-opentype"),
 				/* IE6-IE8 */
-				url("./fonts/basier-square-bold.woff2") format("woff2"),
+				url("/fonts/basier-square-bold.woff2") format("woff2"),
 				/* Super Modern Browsers */
-				url("./fonts/basier-square-bold.woff") format("woff"),
+				url("/fonts/basier-square-bold.woff") format("woff"),
 				/* Modern Browsers */
-				url("./fonts/basier-square-bold.ttf") format("truetype"),
+				url("/fonts/basier-square-bold.ttf") format("truetype"),
 				/* Safari, Android, iOS */
-				url("./fonts/basier-square-bold.svg#Basier Square") format("svg");
+				url("/fonts/basier-square-bold.svg#Basier Square") format("svg");
 			/* Legacy iOS */
 		}
 	</style>
