+++ demo/dapp/src/components/claims/index.ts
@@ -1,6 +1,4 @@
-export {default as Available} from "./Available.svelte";
-export {default as AvailableClaim} from "./AvailableClaim.svelte";
-export {default as Obtained} from "./Obtained.svelte";
-export {default as ObtainedClaim} from "./ObtainedClaim.svelte";
-export {default as SelfSignedForm} from "./SelfSignedForm.svelte";
-export {default as WitnessForm} from "./WitnessForm.svelte";
+export { default as Available } from "./Available.svelte";
+export { default as AvailableClaim } from "./AvailableClaim.svelte";
+export { default as Obtained } from "./Obtained.svelte";
+export { default as ObtainedClaim } from "./ObtainedClaim.svelte";
