+++ demo/dapp/src/components/form/index.ts
@@ -1,1 +1,5 @@
-export { default as CopyTextArea } from './CopyTextArea.svelte'; 
+export { default as CopyTextArea } from './CopyTextArea.svelte';
+export { default as WitnessFormHeader } from './WitnessFormHeader.svelte';
+export { default as WitnessFormStepper } from './WitnessFormStepper.svelte';
+export { default as SelfSignedForm } from './SelfSignedForm.svelte'; 
+export { default as WitnessForm } from "./WitnessForm.svelte";
