+++ demo/dapp/src/components/icons/CloseIcon.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  let clazz: string = '';
+  let clazz: string = "";
   export { clazz as class };
 </script>
