+++ demo/dapp/src/components/icons/DiscordIcon.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import {iconColor as color} from "../../util";
+  import { iconColor as color } from "util";
 </script>
 
 <svg
