+++ demo/dapp/src/components/index.ts
@@ -1,10 +1,7 @@
 export * from './brand/';
 export * from './claims';
 export * from './page/';
-// export * from './typhografy/';
 export * from './buttons/';
-// export * from './select/';
 export * from './icons/';
 export * from './form/';
 export * from './alerts/';
-// export * from './modals/';
